Move fromStateEntityValues to graph state utils
